+++ recipes/linux.rb
@@ -6,7 +6,7 @@
 
 directory '/opt/dynatrace'
 
-remote_file '/opt/dynatrace/Dynatrace-OneAgent-Linux.sh' do
+remote_file "#{Chef::Config[:file_cache_path]}/Dynatrace-OneAgent-Linux.sh" do
   source node['dvo_user']['dynatrace']['linux_installer']
   owner 'root'
   group 'root'
@@ -17,7 +17,7 @@ end
 bash 'Install Dynatrace Agent' do
   cwd '/opt/dynatrace'
   code <<-EOH
-    /bin/sh Dynatrace-OneAgent-Linux.sh APP_LOG_CONTENT_ACCESS=1 #{node['dvo_user']['dynatrace']['infra_only_linux']}
+    /bin/sh #{Chef::Config[:file_cache_path]}/Dynatrace-OneAgent-Linux.sh APP_LOG_CONTENT_ACCESS=1 #{node['dvo_user']['dynatrace']['infra_only_linux']}
     EOH
   not_if { ::File.directory?('/opt/dynatrace/oneagent') }
 end