+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.0
+**Date: 2/16/2018**
+
+* Added uninstall recipe.
+* Fixed windows infrastructure only installation option.
+* Added support for dev tenant.
+
 # 0.0.1
 **Date: 2/2/2018**